Added a modal form dialog box to take user input and save return values as objects in an array.
Could not figure out how to get "dialog.returnValue" to work, so I am using the value of the submit button instead. Seems to work fine. I also have a vision of appending this object to an array and a set.. The array, allBooks, will be the data structure behind the client-side search function. The set, will temporarily hold the book object at the time of creation so as to parse it to a function that injects a div for a card-like display.

TODO: create set for temporary object storage and create function that creates divs for a card-like display that will accept the object stored in the set as a parameter.