FINALLY fixed delete bug
Essentially, added an event listener to every checkbox that updated a Set depending on if that checkboxes' cards' identifier was present in the set or not. In other words, I created a checkbox-triggered dynamically updated set. The identifiers of this set is then compared to the list of identifiers in allCards and if there is a match, the cards associated with those sets are deleted from everywhere (allCards, checkBoxSet, and the DOM). I think i should add this identifier to the book object to clear that as well. Will do it for next commit.